separated boolean controllers
