+++ src/app.js
@@ -30,7 +30,9 @@ const {
   getFirstCharacters,
 } = require('./controllers/strings');
 
-const { negate, truthiness, isOdd, startsWith } = require('./lib/booleans');
+const { negateIt, isTruthy, isOddNum, startsWithChar } = require('./controllers/booleans');
+
+// strings
 
 app.get('/strings/hello/world', practiceHello);
 
@@ -46,6 +48,8 @@ app.get('/strings/count-characters/:string', countingCharacters);
 
 app.get('/strings/first-characters/:string', getFirstCharacters);
 
+// numbers
+
 app.get('/numbers/add/:num1/and/:num2', addNumbers);
 
 app.get('/numbers/subtract/:num2/from/:num1', subtractNumbers);
@@ -55,6 +59,8 @@ app.post('/numbers/multiply', multiplyNumbers);
 app.post('/numbers/divide', divideNumbers);
 
 app.post('/numbers/remainder', remainderOfNumbers);
+
+// arrays
 
 app.post('/arrays/element-at-index/:index', getElement);
 
@@ -66,27 +72,14 @@ app.post('/arrays/starts-with-vowel', startsWithVowel);
 
 app.post('/arrays/remove-element', removeElement);
 
-app.post('/booleans/negate', (req, res) => {
-  res.status(200).send({ result: negate(req.body.value) });
-});
+// booleans
 
-app.post('/booleans/truthiness', (req, res) => {
-  res.status(200).send({ result: truthiness(req.body.value) });
-});
+app.post('/booleans/negate', negateIt);
 
-app.get('/booleans/is-odd/:num', (req, res) => {
-  const number = parseInt(req.params.num, 10);
-  if (Number.isNaN(number)) {
-    return res.status(400).send({ error: 'Parameter must be a number.' });
-  }
-  return res.status(200).send({ result: isOdd(number) });
-});
+app.post('/booleans/truthiness', isTruthy);
 
-app.get('/booleans/:string/starts-with/:char', (req, res) => {
-  if (req.params.char.length > 1) {
-    return res.status(400).send({ error: 'Parameter "character" must be a single character.' });
-  }
-  return res.status(200).send({ result: startsWith(req.params.char, req.params.string) });
-});
+app.get('/booleans/is-odd/:num', isOddNum);
+
+app.get('/booleans/:string/starts-with/:char', startsWithChar);
 
 module.exports = app;
